+++ src/components/batch/CreateBatch.tsx
@@ -5,8 +5,12 @@ import '../global.scss';
 import { DefaultTemplate } from "payload/components/templates";
 import { AdminView } from "payload/config";
 import  formatFields from 'payload/dist/admin/components/views/collections/Edit/formatFields';
+import { Props } from "../types";
 
-const CreateBatch: React.FC = () => {
+
+const CreateBatch: React.FC = (props: Props) => {
+
+    console.log('///CreateBatch Props', props);
 
     return  (
         <section>
