+++ src/components/types.ts
@@ -1,0 +1,5 @@
+import { Props } from 'payload/dist/admin/components/views/collections/Edit/types';
+
+export {
+    Props,
+}
